cast timestamps to gint64 before diffing
git-svn-id: https://delta.affinix.com/svn/trunk/psimedia@524 ac7a513c-5e3a-0410-af42-968c3ac6e7dc

+++ gstprovider/gstelements/speexdsp/speexdsp.c
@@ -272,7 +272,7 @@ pairlog_append_playback (PairLog * pl, const unsigned char * buf, int offset, in
   if (pl->start == GST_CLOCK_TIME_NONE)
     return;
 
-  i = ((time - pl->start) * rate / GST_SECOND) * 2;
+  i = (((gint64)time - (gint64)pl->start) * rate / GST_SECOND) * 2;
   offset = (int)i;
   GST_LOG ("start=%"GST_TIME_FORMAT", time=%"GST_TIME_FORMAT", offset=%d",
       GST_TIME_ARGS (pl->start), GST_TIME_ARGS (time), offset);
@@ -308,7 +308,7 @@ pairlog_append_capture (PairLog * pl, const unsigned char * buf, int offset, int
   if (pl->start == GST_CLOCK_TIME_NONE)
     return;
 
-  i = ((time - pl->start) * rate / GST_SECOND) * 2;
+  i = (((gint64)time - (gint64)pl->start) * rate / GST_SECOND) * 2;
   offset = (int)i;
   GST_LOG ("start=%"GST_TIME_FORMAT", time=%"GST_TIME_FORMAT", offset=%d",
       GST_TIME_ARGS (pl->start), GST_TIME_ARGS (time), offset);
